Change TBAA indexes to Long
Record fields will consume a large number of TBAA indexes. LLVM should
theoretically be fine with 64bit indexes so support them on the Scala
side

+++ compiler/src/main/scala/llambda/codegen/llvmir/IrTbaaNode.scala
@@ -1,6 +1,6 @@
 package llambda.codegen.llvmir
 
-case class IrTbaaNode(index : Int, identity : String, parentIndex : Option[Int] = None, constant : Boolean = false) extends Irable {
+case class IrTbaaNode(index : Long, identity : String, parentIndex : Option[Long] = None, constant : Boolean = false) extends Irable {
   def toIr : String = {
     val identityField = "metadata !\"" + identity + "\""
